xinchao beauty

+++ app/src/main/java/com/example/quizapp/Question.kt
@@ -2,5 +2,5 @@ package com.example.quizapp
 
 data class Question(
     var question: String,
-    var choices: Map<String, Int>
+    var choices: Map<String, Double>
 )
